requirements.txt
django==4.1.3
asgiref==3.5.2
astroid==2.12.13
async-timeout==4.0.2
billiard==3.6.4.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dill==0.3.6
django-environ==0.9.0
django-polymorphic==3.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
factory-boy==3.2.1
Faker==15.3.3
freezegun==1.2.2
geographiclib==2.0
geopy==2.3.0
idna==3.4
isort==5.10.1
Jinja2==3.1.2
kombu==5.2.4
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
mccabe==0.7.0
mock==4.0.3
mysqlclient==2.1.1
packaging==21.3
pdfkit==1.0.0
platformdirs==2.5.4
prompt-toolkit==3.0.33
psycopg2-binary==2.9.5
PyJWT==2.6.0
pylint==2.15.6
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
responses==0.22.0
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
types-toml==0.10.8.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
wkhtmltopdf==0.2
wrapt==1.14.1
redis==4.3.5
